+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-# Edge
+# 12.10.0
 
 - Microcosm ships with ES6 and UMD bundles
 - Domains and Effects can implement a `defaults` static object to
   provide default setup options.
 - Do not return undefined when using `get` to retrieve a null value
   without a fallback.
+- `repo.append(action, state)` should reconcile history
 
 # 12.9.0